+++ src/put/zips.rs
@@ -8,13 +8,13 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Serialize, Deserialize)]
 pub struct CreateZipResponse {
-    pub zip_id: u32,
+    pub zip_id: i64,
 }
 
 /// Creates a new ZIP job with the given file id.
 ///
 /// Waits for the zip job complete, and returns a string with the download URL.
-pub fn create(client: &Client, api_token: &String, file_id: u32) -> Result<String, Error> {
+pub fn create(client: &Client, api_token: &String, file_id: i64) -> Result<String, Error> {
     // Start ZIP job
     let form: Form = Form::new().text("file_ids", file_id.to_string());
 
@@ -45,7 +45,7 @@ pub struct CheckZipResponse {
 }
 
 /// Checks the status of a given zip job
-pub fn get(client: &Client, api_token: &String, zip_id: u32) -> Result<CheckZipResponse, Error> {
+pub fn get(client: &Client, api_token: &String, zip_id: i64) -> Result<CheckZipResponse, Error> {
     let response: CheckZipResponse = client
         .get(format!("https://api.put.io/v2/zips/{zip_id}"))
         .header("authorization", format!("Bearer {api_token}"))
